+++ lib/gamefic/scene.rb
@@ -8,6 +8,7 @@ module Gamefic
     require 'gamefic/scene/default'
     require 'gamefic/scene/activity'
     require 'gamefic/scene/multiple_choice'
+    require 'gamefic/scene/active_choice'
     require 'gamefic/scene/pause'
     require 'gamefic/scene/yes_or_no'
     require 'gamefic/scene/conclusion'