+++ lib/gamefic/scanner/result.rb
@@ -38,10 +38,6 @@ module Gamefic
         @processor = processor
       end
 
-      def passed
-        @passed ||= (token.keywords - remainder.keywords).join(' ')
-      end
-
       # The strictness of the scanner that produced this result.
       #
       # @return [Integer]